коррекция количества итемов при создании чара

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3007,7 +3007,7 @@ void ObjectMgr::LoadPlayerInfo()
                     continue;
                 }
 
-                int32 amount   = fields[3].GetInt8();
+                int32 amount   = fields[3].GetUInt32();
 
                 if (!amount)
                 {
